LayoutTests/fast/mediastream/peerconnection-Attributes-expected.txt

Tests PeerConnection00 related attributes according to http://www.w3.org/TR/webrtc/

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof pc.addStream === 'function' is true
PASS typeof pc.removeStream === 'function' is true
PASS typeof pc.close === 'function' is true
PASS pc.NEW === 0 is true
PASS pc.OPENING === 1 is true
PASS pc.ACTIVE === 2 is true
PASS pc.CLOSED === 3 is true
PASS pc.ICE_GATHERING === 0x100 is true
PASS pc.ICE_WAITING === 0x200 is true
PASS pc.ICE_CHECKING === 0x300 is true
PASS pc.ICE_CONNECTED === 0x400 is true
PASS pc.ICE_COMPLETED === 0x500 is true
PASS pc.ICE_FAILED === 0x600 is true
PASS pc.ICE_CLOSED === 0x700 is true
PASS pc.SDP_OFFER === 0x100 is true
PASS pc.SDP_PRANSWER === 0x200 is true
PASS pc.SDP_ANSWER ===  0x300 is true
PASS typeof pc.localStreams === 'object' is true
PASS typeof pc.remoteStreams === 'object' is true
PASS typeof pc.onaddstream === 'object' is true
PASS typeof pc.onremovestream === 'object' is true
PASS typeof pc.onconnecting === 'object' is true
PASS typeof pc.onopen === 'object' is true
PASS successfullyParsed is true

TEST COMPLETE
